+++ adb-20190315/alibabacloud_adb20190315/models.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 # This file is auto-generated, don't edit it. Thanks.
 from Tea.model import TeaModel
-from typing import Dict, List
+from typing import Dict, List, Any
 
 
 class AllocateClusterPublicConnectionRequest(TeaModel):
@@ -1923,9 +1923,17 @@ class CreateDBResourceGroupRequest(TeaModel):
     def __init__(
         self,
         client_token: str = None,
+        cluster_mode: str = None,
+        cluster_size_resource: str = None,
         dbcluster_id: str = None,
+        engine: str = None,
+        engine_params: Dict[str, Any] = None,
         group_name: str = None,
         group_type: str = None,
+        max_cluster_count: int = None,
+        max_compute_resource: str = None,
+        min_cluster_count: int = None,
+        min_compute_resource: str = None,
         node_num: int = None,
         owner_account: str = None,
         owner_id: int = None,
@@ -1933,12 +1941,16 @@ class CreateDBResourceGroupRequest(TeaModel):
         resource_owner_id: int = None,
     ):
         self.client_token = client_token
+        self.cluster_mode = cluster_mode
+        self.cluster_size_resource = cluster_size_resource
         # The ID of the AnalyticDB for MySQL Data Warehouse Edition (V3.0) cluster.
         # 
         # >  You can call the [DescribeDBClusters](https://help.aliyun.com/document_detail/129857.html) operation to query the cluster IDs of all AnalyticDB for MySQL Data Warehouse Edition (V3.0) clusters within a specific region.
         # 
         # This parameter is required.
         self.dbcluster_id = dbcluster_id
+        self.engine = engine
+        self.engine_params = engine_params
         # The name of the resource group.
         # 
         # *   The name can be up to 255 characters in length.
@@ -1952,6 +1964,10 @@ class CreateDBResourceGroupRequest(TeaModel):
         # *   **interactive** (default)
         # *   **batch**\
         self.group_type = group_type
+        self.max_cluster_count = max_cluster_count
+        self.max_compute_resource = max_compute_resource
+        self.min_cluster_count = min_cluster_count
+        self.min_compute_resource = min_compute_resource
         # The number of nodes. Default value: 0.
         # 
         # *   Each node is configured with the resources of 16 cores and 64 GB memory.
@@ -1973,12 +1989,28 @@ class CreateDBResourceGroupRequest(TeaModel):
         result = dict()
         if self.client_token is not None:
             result['ClientToken'] = self.client_token
+        if self.cluster_mode is not None:
+            result['ClusterMode'] = self.cluster_mode
+        if self.cluster_size_resource is not None:
+            result['ClusterSizeResource'] = self.cluster_size_resource
         if self.dbcluster_id is not None:
             result['DBClusterId'] = self.dbcluster_id
+        if self.engine is not None:
+            result['Engine'] = self.engine
+        if self.engine_params is not None:
+            result['EngineParams'] = self.engine_params
         if self.group_name is not None:
             result['GroupName'] = self.group_name
         if self.group_type is not None:
             result['GroupType'] = self.group_type
+        if self.max_cluster_count is not None:
+            result['MaxClusterCount'] = self.max_cluster_count
+        if self.max_compute_resource is not None:
+            result['MaxComputeResource'] = self.max_compute_resource
+        if self.min_cluster_count is not None:
+            result['MinClusterCount'] = self.min_cluster_count
+        if self.min_compute_resource is not None:
+            result['MinComputeResource'] = self.min_compute_resource
         if self.node_num is not None:
             result['NodeNum'] = self.node_num
         if self.owner_account is not None:
@@ -1995,12 +2027,171 @@ class CreateDBResourceGroupRequest(TeaModel):
         m = m or dict()
         if m.get('ClientToken') is not None:
             self.client_token = m.get('ClientToken')
+        if m.get('ClusterMode') is not None:
+            self.cluster_mode = m.get('ClusterMode')
+        if m.get('ClusterSizeResource') is not None:
+            self.cluster_size_resource = m.get('ClusterSizeResource')
         if m.get('DBClusterId') is not None:
             self.dbcluster_id = m.get('DBClusterId')
+        if m.get('Engine') is not None:
+            self.engine = m.get('Engine')
+        if m.get('EngineParams') is not None:
+            self.engine_params = m.get('EngineParams')
         if m.get('GroupName') is not None:
             self.group_name = m.get('GroupName')
         if m.get('GroupType') is not None:
             self.group_type = m.get('GroupType')
+        if m.get('MaxClusterCount') is not None:
+            self.max_cluster_count = m.get('MaxClusterCount')
+        if m.get('MaxComputeResource') is not None:
+            self.max_compute_resource = m.get('MaxComputeResource')
+        if m.get('MinClusterCount') is not None:
+            self.min_cluster_count = m.get('MinClusterCount')
+        if m.get('MinComputeResource') is not None:
+            self.min_compute_resource = m.get('MinComputeResource')
+        if m.get('NodeNum') is not None:
+            self.node_num = m.get('NodeNum')
+        if m.get('OwnerAccount') is not None:
+            self.owner_account = m.get('OwnerAccount')
+        if m.get('OwnerId') is not None:
+            self.owner_id = m.get('OwnerId')
+        if m.get('ResourceOwnerAccount') is not None:
+            self.resource_owner_account = m.get('ResourceOwnerAccount')
+        if m.get('ResourceOwnerId') is not None:
+            self.resource_owner_id = m.get('ResourceOwnerId')
+        return self
+
+
+class CreateDBResourceGroupShrinkRequest(TeaModel):
+    def __init__(
+        self,
+        client_token: str = None,
+        cluster_mode: str = None,
+        cluster_size_resource: str = None,
+        dbcluster_id: str = None,
+        engine: str = None,
+        engine_params_shrink: str = None,
+        group_name: str = None,
+        group_type: str = None,
+        max_cluster_count: int = None,
+        max_compute_resource: str = None,
+        min_cluster_count: int = None,
+        min_compute_resource: str = None,
+        node_num: int = None,
+        owner_account: str = None,
+        owner_id: int = None,
+        resource_owner_account: str = None,
+        resource_owner_id: int = None,
+    ):
+        self.client_token = client_token
+        self.cluster_mode = cluster_mode
+        self.cluster_size_resource = cluster_size_resource
+        # The ID of the AnalyticDB for MySQL Data Warehouse Edition (V3.0) cluster.
+        # 
+        # >  You can call the [DescribeDBClusters](https://help.aliyun.com/document_detail/129857.html) operation to query the cluster IDs of all AnalyticDB for MySQL Data Warehouse Edition (V3.0) clusters within a specific region.
+        # 
+        # This parameter is required.
+        self.dbcluster_id = dbcluster_id
+        self.engine = engine
+        self.engine_params_shrink = engine_params_shrink
+        # The name of the resource group.
+        # 
+        # *   The name can be up to 255 characters in length.
+        # *   The name must start with an uppercase letter or a digit.
+        # *   The name can contain uppercase letters, digits, hyphens (-), and underscores (_).
+        # 
+        # This parameter is required.
+        self.group_name = group_name
+        # The query execution mode. Valid values:
+        # 
+        # *   **interactive** (default)
+        # *   **batch**\
+        self.group_type = group_type
+        self.max_cluster_count = max_cluster_count
+        self.max_compute_resource = max_compute_resource
+        self.min_cluster_count = min_cluster_count
+        self.min_compute_resource = min_compute_resource
+        # The number of nodes. Default value: 0.
+        # 
+        # *   Each node is configured with the resources of 16 cores and 64 GB memory.
+        # *   Make sure that the amount of resources of the nodes (Number of nodes × 16 cores and 64 GB memory) is less than or equal to the amount of unused resources of the cluster.
+        self.node_num = node_num
+        self.owner_account = owner_account
+        self.owner_id = owner_id
+        self.resource_owner_account = resource_owner_account
+        self.resource_owner_id = resource_owner_id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.client_token is not None:
+            result['ClientToken'] = self.client_token
+        if self.cluster_mode is not None:
+            result['ClusterMode'] = self.cluster_mode
+        if self.cluster_size_resource is not None:
+            result['ClusterSizeResource'] = self.cluster_size_resource
+        if self.dbcluster_id is not None:
+            result['DBClusterId'] = self.dbcluster_id
+        if self.engine is not None:
+            result['Engine'] = self.engine
+        if self.engine_params_shrink is not None:
+            result['EngineParams'] = self.engine_params_shrink
+        if self.group_name is not None:
+            result['GroupName'] = self.group_name
+        if self.group_type is not None:
+            result['GroupType'] = self.group_type
+        if self.max_cluster_count is not None:
+            result['MaxClusterCount'] = self.max_cluster_count
+        if self.max_compute_resource is not None:
+            result['MaxComputeResource'] = self.max_compute_resource
+        if self.min_cluster_count is not None:
+            result['MinClusterCount'] = self.min_cluster_count
+        if self.min_compute_resource is not None:
+            result['MinComputeResource'] = self.min_compute_resource
+        if self.node_num is not None:
+            result['NodeNum'] = self.node_num
+        if self.owner_account is not None:
+            result['OwnerAccount'] = self.owner_account
+        if self.owner_id is not None:
+            result['OwnerId'] = self.owner_id
+        if self.resource_owner_account is not None:
+            result['ResourceOwnerAccount'] = self.resource_owner_account
+        if self.resource_owner_id is not None:
+            result['ResourceOwnerId'] = self.resource_owner_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('ClientToken') is not None:
+            self.client_token = m.get('ClientToken')
+        if m.get('ClusterMode') is not None:
+            self.cluster_mode = m.get('ClusterMode')
+        if m.get('ClusterSizeResource') is not None:
+            self.cluster_size_resource = m.get('ClusterSizeResource')
+        if m.get('DBClusterId') is not None:
+            self.dbcluster_id = m.get('DBClusterId')
+        if m.get('Engine') is not None:
+            self.engine = m.get('Engine')
+        if m.get('EngineParams') is not None:
+            self.engine_params_shrink = m.get('EngineParams')
+        if m.get('GroupName') is not None:
+            self.group_name = m.get('GroupName')
+        if m.get('GroupType') is not None:
+            self.group_type = m.get('GroupType')
+        if m.get('MaxClusterCount') is not None:
+            self.max_cluster_count = m.get('MaxClusterCount')
+        if m.get('MaxComputeResource') is not None:
+            self.max_compute_resource = m.get('MaxComputeResource')
+        if m.get('MinClusterCount') is not None:
+            self.min_cluster_count = m.get('MinClusterCount')
+        if m.get('MinComputeResource') is not None:
+            self.min_compute_resource = m.get('MinComputeResource')
         if m.get('NodeNum') is not None:
             self.node_num = m.get('NodeNum')
         if m.get('OwnerAccount') is not None:
@@ -11129,6 +11320,7 @@ class DescribeDBClusterNetInfoRequest(TeaModel):
     def __init__(
         self,
         dbcluster_id: str = None,
+        engine: str = None,
         owner_account: str = None,
         owner_id: int = None,
         region_id: str = None,
@@ -11139,6 +11331,7 @@ class DescribeDBClusterNetInfoRequest(TeaModel):
         # 
         # This parameter is required.
         self.dbcluster_id = dbcluster_id
+        self.engine = engine
         self.owner_account = owner_account
         self.owner_id = owner_id
         # The region ID of the cluster.
@@ -11157,6 +11350,8 @@ class DescribeDBClusterNetInfoRequest(TeaModel):
         result = dict()
         if self.dbcluster_id is not None:
             result['DBClusterId'] = self.dbcluster_id
+        if self.engine is not None:
+            result['Engine'] = self.engine
         if self.owner_account is not None:
             result['OwnerAccount'] = self.owner_account
         if self.owner_id is not None:
@@ -11173,6 +11368,8 @@ class DescribeDBClusterNetInfoRequest(TeaModel):
         m = m or dict()
         if m.get('DBClusterId') is not None:
             self.dbcluster_id = m.get('DBClusterId')
+        if m.get('Engine') is not None:
+            self.engine = m.get('Engine')
         if m.get('OwnerAccount') is not None:
             self.owner_account = m.get('OwnerAccount')
         if m.get('OwnerId') is not None:
@@ -13812,16 +14009,35 @@ class DescribeDBResourceGroupRequest(TeaModel):
 class DescribeDBResourceGroupResponseBodyGroupsInfo(TeaModel):
     def __init__(
         self,
+        cluster_mode: str = None,
+        cluster_size_resource: str = None,
+        connection_string: str = None,
         create_time: str = None,
+        elastic_min_compute_resource: str = None,
+        engine: str = None,
+        engine_params: Dict[str, Any] = None,
         group_name: str = None,
         group_type: str = None,
         group_user_list: List[str] = None,
         group_users: str = None,
+        max_cluster_count: int = None,
+        max_compute_resource: str = None,
+        min_cluster_count: int = None,
+        min_compute_resource: str = None,
         node_num: int = None,
+        port: str = None,
+        running_cluster_count: int = None,
+        status: str = None,
         update_time: str = None,
     ):
+        self.cluster_mode = cluster_mode
+        self.cluster_size_resource = cluster_size_resource
+        self.connection_string = connection_string
         # The time when the resource group was created.
         self.create_time = create_time
+        self.elastic_min_compute_resource = elastic_min_compute_resource
+        self.engine = engine
+        self.engine_params = engine_params
         # The name of the resource group.
         self.group_name = group_name
         # The query execution mode. Valid values:
@@ -13835,8 +14051,15 @@ class DescribeDBResourceGroupResponseBodyGroupsInfo(TeaModel):
         self.group_user_list = group_user_list
         # The database accounts that are associated with the resource group. Multiple database accounts are separated by commas (,).
         self.group_users = group_users
+        self.max_cluster_count = max_cluster_count
+        self.max_compute_resource = max_compute_resource
+        self.min_cluster_count = min_cluster_count
+        self.min_compute_resource = min_compute_resource
         # The number of nodes. Each node provides 16 cores and 64 GB memory.
         self.node_num = node_num
+        self.port = port
+        self.running_cluster_count = running_cluster_count
+        self.status = status
         # The time when the resource group was updated.
         self.update_time = update_time
 
@@ -13849,8 +14072,20 @@ class DescribeDBResourceGroupResponseBodyGroupsInfo(TeaModel):
             return _map
 
         result = dict()
+        if self.cluster_mode is not None:
+            result['ClusterMode'] = self.cluster_mode
+        if self.cluster_size_resource is not None:
+            result['ClusterSizeResource'] = self.cluster_size_resource
+        if self.connection_string is not None:
+            result['ConnectionString'] = self.connection_string
         if self.create_time is not None:
             result['CreateTime'] = self.create_time
+        if self.elastic_min_compute_resource is not None:
+            result['ElasticMinComputeResource'] = self.elastic_min_compute_resource
+        if self.engine is not None:
+            result['Engine'] = self.engine
+        if self.engine_params is not None:
+            result['EngineParams'] = self.engine_params
         if self.group_name is not None:
             result['GroupName'] = self.group_name
         if self.group_type is not None:
@@ -13859,16 +14094,42 @@ class DescribeDBResourceGroupResponseBodyGroupsInfo(TeaModel):
             result['GroupUserList'] = self.group_user_list
         if self.group_users is not None:
             result['GroupUsers'] = self.group_users
+        if self.max_cluster_count is not None:
+            result['MaxClusterCount'] = self.max_cluster_count
+        if self.max_compute_resource is not None:
+            result['MaxComputeResource'] = self.max_compute_resource
+        if self.min_cluster_count is not None:
+            result['MinClusterCount'] = self.min_cluster_count
+        if self.min_compute_resource is not None:
+            result['MinComputeResource'] = self.min_compute_resource
         if self.node_num is not None:
             result['NodeNum'] = self.node_num
+        if self.port is not None:
+            result['Port'] = self.port
+        if self.running_cluster_count is not None:
+            result['RunningClusterCount'] = self.running_cluster_count
+        if self.status is not None:
+            result['Status'] = self.status
         if self.update_time is not None:
             result['UpdateTime'] = self.update_time
         return result
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('ClusterMode') is not None:
+            self.cluster_mode = m.get('ClusterMode')
+        if m.get('ClusterSizeResource') is not None:
+            self.cluster_size_resource = m.get('ClusterSizeResource')
+        if m.get('ConnectionString') is not None:
+            self.connection_string = m.get('ConnectionString')
         if m.get('CreateTime') is not None:
             self.create_time = m.get('CreateTime')
+        if m.get('ElasticMinComputeResource') is not None:
+            self.elastic_min_compute_resource = m.get('ElasticMinComputeResource')
+        if m.get('Engine') is not None:
+            self.engine = m.get('Engine')
+        if m.get('EngineParams') is not None:
+            self.engine_params = m.get('EngineParams')
         if m.get('GroupName') is not None:
             self.group_name = m.get('GroupName')
         if m.get('GroupType') is not None:
@@ -13877,8 +14138,22 @@ class DescribeDBResourceGroupResponseBodyGroupsInfo(TeaModel):
             self.group_user_list = m.get('GroupUserList')
         if m.get('GroupUsers') is not None:
             self.group_users = m.get('GroupUsers')
+        if m.get('MaxClusterCount') is not None:
+            self.max_cluster_count = m.get('MaxClusterCount')
+        if m.get('MaxComputeResource') is not None:
+            self.max_compute_resource = m.get('MaxComputeResource')
+        if m.get('MinClusterCount') is not None:
+            self.min_cluster_count = m.get('MinClusterCount')
+        if m.get('MinComputeResource') is not None:
+            self.min_compute_resource = m.get('MinComputeResource')
         if m.get('NodeNum') is not None:
             self.node_num = m.get('NodeNum')
+        if m.get('Port') is not None:
+            self.port = m.get('Port')
+        if m.get('RunningClusterCount') is not None:
+            self.running_cluster_count = m.get('RunningClusterCount')
+        if m.get('Status') is not None:
+            self.status = m.get('Status')
         if m.get('UpdateTime') is not None:
             self.update_time = m.get('UpdateTime')
         return self
@@ -33920,9 +34195,16 @@ class ModifyDBResourceGroupRequest(TeaModel):
     def __init__(
         self,
         client_token: str = None,
+        cluster_mode: str = None,
+        cluster_size_resource: str = None,
         dbcluster_id: str = None,
+        engine_params: Dict[str, Any] = None,
         group_name: str = None,
         group_type: str = None,
+        max_cluster_count: int = None,
+        max_compute_resource: str = None,
+        min_cluster_count: int = None,
+        min_compute_resource: str = None,
         node_num: int = None,
         owner_account: str = None,
         owner_id: int = None,
@@ -33932,12 +34214,15 @@ class ModifyDBResourceGroupRequest(TeaModel):
     ):
         # The client token that is used to ensure the idempotence of the request. You can use the client to generate the token, but you must make sure that the token is unique among different requests. The token can contain only ASCII characters and cannot exceed 64 characters in length.
         self.client_token = client_token
+        self.cluster_mode = cluster_mode
+        self.cluster_size_resource = cluster_size_resource
         # The ID of the AnalyticDB for MySQL Data Warehouse Edition (V3.0) cluster.
         # 
         # >  You can call the [DescribeDBClusters](https://help.aliyun.com/document_detail/129857.html) operation to query the cluster IDs of all AnalyticDB for MySQL Data Warehouse Edition (V3.0) clusters within a specific region.
         # 
         # This parameter is required.
         self.dbcluster_id = dbcluster_id
+        self.engine_params = engine_params
         # The name of the resource group.
         # 
         # This parameter is required.
@@ -33949,6 +34234,10 @@ class ModifyDBResourceGroupRequest(TeaModel):
         # 
         # >  For more information, see [Query execution modes](https://help.aliyun.com/document_detail/189502.html).
         self.group_type = group_type
+        self.max_cluster_count = max_cluster_count
+        self.max_compute_resource = max_compute_resource
+        self.min_cluster_count = min_cluster_count
+        self.min_compute_resource = min_compute_resource
         # The number of nodes. Default value: 0.
         # 
         # *   Each node is configured with the resources of 16 cores and 64 GB memory.
@@ -33972,12 +34261,26 @@ class ModifyDBResourceGroupRequest(TeaModel):
         result = dict()
         if self.client_token is not None:
             result['ClientToken'] = self.client_token
+        if self.cluster_mode is not None:
+            result['ClusterMode'] = self.cluster_mode
+        if self.cluster_size_resource is not None:
+            result['ClusterSizeResource'] = self.cluster_size_resource
         if self.dbcluster_id is not None:
             result['DBClusterId'] = self.dbcluster_id
+        if self.engine_params is not None:
+            result['EngineParams'] = self.engine_params
         if self.group_name is not None:
             result['GroupName'] = self.group_name
         if self.group_type is not None:
             result['GroupType'] = self.group_type
+        if self.max_cluster_count is not None:
+            result['MaxClusterCount'] = self.max_cluster_count
+        if self.max_compute_resource is not None:
+            result['MaxComputeResource'] = self.max_compute_resource
+        if self.min_cluster_count is not None:
+            result['MinClusterCount'] = self.min_cluster_count
+        if self.min_compute_resource is not None:
+            result['MinComputeResource'] = self.min_compute_resource
         if self.node_num is not None:
             result['NodeNum'] = self.node_num
         if self.owner_account is not None:
@@ -33996,12 +34299,26 @@ class ModifyDBResourceGroupRequest(TeaModel):
         m = m or dict()
         if m.get('ClientToken') is not None:
             self.client_token = m.get('ClientToken')
+        if m.get('ClusterMode') is not None:
+            self.cluster_mode = m.get('ClusterMode')
+        if m.get('ClusterSizeResource') is not None:
+            self.cluster_size_resource = m.get('ClusterSizeResource')
         if m.get('DBClusterId') is not None:
             self.dbcluster_id = m.get('DBClusterId')
+        if m.get('EngineParams') is not None:
+            self.engine_params = m.get('EngineParams')
         if m.get('GroupName') is not None:
             self.group_name = m.get('GroupName')
         if m.get('GroupType') is not None:
             self.group_type = m.get('GroupType')
+        if m.get('MaxClusterCount') is not None:
+            self.max_cluster_count = m.get('MaxClusterCount')
+        if m.get('MaxComputeResource') is not None:
+            self.max_compute_resource = m.get('MaxComputeResource')
+        if m.get('MinClusterCount') is not None:
+            self.min_cluster_count = m.get('MinClusterCount')
+        if m.get('MinComputeResource') is not None:
+            self.min_compute_resource = m.get('MinComputeResource')
         if m.get('NodeNum') is not None:
             self.node_num = m.get('NodeNum')
         if m.get('OwnerAccount') is not None:
@@ -34021,9 +34338,16 @@ class ModifyDBResourceGroupShrinkRequest(TeaModel):
     def __init__(
         self,
         client_token: str = None,
+        cluster_mode: str = None,
+        cluster_size_resource: str = None,
         dbcluster_id: str = None,
+        engine_params_shrink: str = None,
         group_name: str = None,
         group_type: str = None,
+        max_cluster_count: int = None,
+        max_compute_resource: str = None,
+        min_cluster_count: int = None,
+        min_compute_resource: str = None,
         node_num: int = None,
         owner_account: str = None,
         owner_id: int = None,
@@ -34033,12 +34357,15 @@ class ModifyDBResourceGroupShrinkRequest(TeaModel):
     ):
         # The client token that is used to ensure the idempotence of the request. You can use the client to generate the token, but you must make sure that the token is unique among different requests. The token can contain only ASCII characters and cannot exceed 64 characters in length.
         self.client_token = client_token
+        self.cluster_mode = cluster_mode
+        self.cluster_size_resource = cluster_size_resource
         # The ID of the AnalyticDB for MySQL Data Warehouse Edition (V3.0) cluster.
         # 
         # >  You can call the [DescribeDBClusters](https://help.aliyun.com/document_detail/129857.html) operation to query the cluster IDs of all AnalyticDB for MySQL Data Warehouse Edition (V3.0) clusters within a specific region.
         # 
         # This parameter is required.
         self.dbcluster_id = dbcluster_id
+        self.engine_params_shrink = engine_params_shrink
         # The name of the resource group.
         # 
         # This parameter is required.
@@ -34050,6 +34377,10 @@ class ModifyDBResourceGroupShrinkRequest(TeaModel):
         # 
         # >  For more information, see [Query execution modes](https://help.aliyun.com/document_detail/189502.html).
         self.group_type = group_type
+        self.max_cluster_count = max_cluster_count
+        self.max_compute_resource = max_compute_resource
+        self.min_cluster_count = min_cluster_count
+        self.min_compute_resource = min_compute_resource
         # The number of nodes. Default value: 0.
         # 
         # *   Each node is configured with the resources of 16 cores and 64 GB memory.
@@ -34073,12 +34404,26 @@ class ModifyDBResourceGroupShrinkRequest(TeaModel):
         result = dict()
         if self.client_token is not None:
             result['ClientToken'] = self.client_token
+        if self.cluster_mode is not None:
+            result['ClusterMode'] = self.cluster_mode
+        if self.cluster_size_resource is not None:
+            result['ClusterSizeResource'] = self.cluster_size_resource
         if self.dbcluster_id is not None:
             result['DBClusterId'] = self.dbcluster_id
+        if self.engine_params_shrink is not None:
+            result['EngineParams'] = self.engine_params_shrink
         if self.group_name is not None:
             result['GroupName'] = self.group_name
         if self.group_type is not None:
             result['GroupType'] = self.group_type
+        if self.max_cluster_count is not None:
+            result['MaxClusterCount'] = self.max_cluster_count
+        if self.max_compute_resource is not None:
+            result['MaxComputeResource'] = self.max_compute_resource
+        if self.min_cluster_count is not None:
+            result['MinClusterCount'] = self.min_cluster_count
+        if self.min_compute_resource is not None:
+            result['MinComputeResource'] = self.min_compute_resource
         if self.node_num is not None:
             result['NodeNum'] = self.node_num
         if self.owner_account is not None:
@@ -34097,12 +34442,26 @@ class ModifyDBResourceGroupShrinkRequest(TeaModel):
         m = m or dict()
         if m.get('ClientToken') is not None:
             self.client_token = m.get('ClientToken')
+        if m.get('ClusterMode') is not None:
+            self.cluster_mode = m.get('ClusterMode')
+        if m.get('ClusterSizeResource') is not None:
+            self.cluster_size_resource = m.get('ClusterSizeResource')
         if m.get('DBClusterId') is not None:
             self.dbcluster_id = m.get('DBClusterId')
+        if m.get('EngineParams') is not None:
+            self.engine_params_shrink = m.get('EngineParams')
         if m.get('GroupName') is not None:
             self.group_name = m.get('GroupName')
         if m.get('GroupType') is not None:
             self.group_type = m.get('GroupType')
+        if m.get('MaxClusterCount') is not None:
+            self.max_cluster_count = m.get('MaxClusterCount')
+        if m.get('MaxComputeResource') is not None:
+            self.max_compute_resource = m.get('MaxComputeResource')
+        if m.get('MinClusterCount') is not None:
+            self.min_cluster_count = m.get('MinClusterCount')
+        if m.get('MinComputeResource') is not None:
+            self.min_compute_resource = m.get('MinComputeResource')
         if m.get('NodeNum') is not None:
             self.node_num = m.get('NodeNum')
         if m.get('OwnerAccount') is not None: